Fixed Authentication & Authorization for Users

+++ app/Http/Controllers/ThreadController.php
@@ -7,6 +7,10 @@ use Illuminate\Http\Request;
 
 class ThreadController extends Controller
 {
+    function __construct()
+    {
+        return $this->middleware('auth')->except('index');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -45,7 +49,7 @@ class ThreadController extends Controller
         ]);
 
         //store
-        Thread::create($request->all());
+        $thread = auth()->user()->threads()->create($request->all());
 
         //redirect
         return redirect()->back()->with('message', 'Threat Created');
@@ -82,6 +86,10 @@ class ThreadController extends Controller
      */
     public function update(Request $request, Thread $thread)
     {
+        if(auth()->user()->id !== $thread->user_id){
+            abort(401,"unauthorized");
+        }
+        
         //validate
         $this->validate($request, [
             'subject' => 'required|min:5',
@@ -104,6 +112,10 @@ class ThreadController extends Controller
      */
     public function destroy(Thread $thread)
     {
+        if(auth()->user()->id !== $thread->user_id){
+            abort(401,"unauthorized");
+        }
+
         $thread->delete();
 
         return redirect()->route('thread.index')->with('message', 'Threat Deleted');
